requirements.txt
gevent==1.4.0
loguru==0.5.3
Python==3.7.6
numpy==1.18.1
Pillow==7.0.0
torch==1.6.0
terminaltables==3.1.0
torchvision==0.6.0
tqdm==4.42.1